Order read and manage functions
Been using a lot of recursion for these. I learnt that it was a thing from this video: https://www.youtube.com/watch?v=nuML9SmdbJ4
I mainly use it it cause it reduces nesting in the code. It feels simpler.
Results in more functions though, and my naming conventions for these aren't really making sense.