proto script
So yeah, system params; I discovered these from a project that I was developing in a jam-style fashion, during that I introduced myself to system params.
So these things seem to replace the callback processing architecture that I had been previously building. It's quicker to create, is less bug-prone, and is ergonomically the same; Amazing.
So yeah I feel like I've discovered a way I should/are-supposed-to do things in Bevy with this.

I will likely link the other project at some point, but not right now. Eventually though yeah, because I learned quite a lot from it, and I want to write about that.