Movement system reverted
Reverted to the architecture I used before introducing traits.
Bevy doesn't support querying for traits out of the box. There is crate that lets you do this though:
https://crates.io/crates/bevy-trait-query
I decided to just keep it simple and not use that, not that it looks complicated to use.

Naming convention for MoveToMover is to be changed.