Detector rework finished
That is it mostly done, but there are some things that I can't tackle yet, so I can't complete it.
I need to have the stuff for handling attackable and detectable so I can have a struct that describes an Entity that is holding a unit's 'soul'.
I've called it soul, it is the stuff that is able to be detected and attacked. If a unit didn't have these, they'd exist but would be kind of ghost-like in that they can't be detected or attacked.