Pre rework detector changes
There's also a renaming of a ting in there.
Low on brain-power rn.

It's not a major rework, I just need to add some other stuff.
Right now, or I've made some changes so it isn't fully like that, but right now, it queries for all the outputs in the main detection thing.

The thing is, this detector will be used by units that aren't going to have target unit detection.
So the detector needs to automatically adapt so that it only outputs to the detection components that are there (and that it has implemented).

Btw
Detector is the thing that detects, detection is a type of output from a detector. They both exist as components.

So I can do this, I know how to do it. It'll just have to store the outputs in itself, then have separate systems for outputting to the detection components. That way, those systems only get called if the combo of this component plus the detection component exists. Brain power all used up though, hungry.

I also need to do a component that'll take the target from a commandable component, and output it to the detection component. The detector also needs to store the target and do the same isolated query stuff to get the target itself, so that the main update doesn't have to query for the thing.

Basically, all the conditional outputs and inputs need their own dedicated systems.

This does make me appreciate Bevy more, that I can do it like this.