Control block
Plus integrated block
Which lets me spawn a thing, and have the things that point to it, in the root, spawn with that function call.

Also, the previous commit has some other changes mixed into it. Most if it was just moving around files.
However, notably, I created a macro for automating the impls for the entity components that the different unit entities use.
It shouldn't have been all bunched like that in one commit, but it'd be a waste of time to actually deal with that.