Entity reference flag
Okay so this is my solution.
A query stack doesn't work/fit in bevy. The system is the query stack.
I could still create an implementation that binds a struct for the system params to the reference flag. I don't think that makes sense though.
The reference flag merely shows that the reference exists, and the entity pathway to get there. It would be used in the editor, to show dependencies that are needed, or as a error log feature.
The system params, would just show the systems used by the system that the component flags. It isn't helpful in editor, and it doesn't automate any of the process for creating the system.

I suppose, there could be value in having another implementation that shows all the queries that use a component, so a developer could know more about how something should be set up.
I would make that separate from this.

Hmm...
With the previous implementation the dependency was implicit in that it had a field for an entity. Here, it travels along the pathways I have pre-defined. In one sense it is more tidy, it means that you don't get a web of dependencies. In another sense, the dependency is more obfuscated. With the right editor functionality though, it would show it properly.

Also yeah, the trait would be an editor only thing. Hmm... I don't know if it is possible to have a conditional trait or implementation like that though. The component itself is runtime and editor, the reference flag implementation is only relevant to the editor.

Hmm... With compiler flags it is probably possible, but I'm not sure if the rest of the Bevy ecosystem does this. It makes me think that this is a misuse of bevy.

I suppose the solution then would be another component that has this functionality. Then there would have to be something that couples it to the runtime component though. Seems like you'd run into the same issue to me.

You could do a bundle. In the imagined editor though, developers would be able to spawn components specifically.

Hmm... I suppose the editor itself could maybe detect the components and then spawn these with them? Then it'd have to be that these components are also flagged to not be shown in the inspector. Hmm...

How would it detect these aswell? You have to couple it somehow. Would you do it in the spawning of the component? That means that you'd have to define special conditions for certain component spawns. If it is one of these per entity aswell, then the flag itself would have to become responsible for all things that utallise it on that entity.

Hmm... I think these editor components should still be inspectable too, just separated from normal ones.

Okay, no. Ideally the implementation would just be conditional, I think that can be done via a derive macro. Then I guess you'd have that in a crate that only gets compiled for edit mode. Though, what happens to the macro when runtime is compiled? Okay so the trait would be in an editor only crate; the macro would have to be in something that is runtime compiled.

I wonder if this is just a Unity brain thing or not? It doesn't seem like it'd be bad to me; I suppose I should take a look at how some of these editors work though.