Old entity refs deleted
These were replaced by the ones declared in the result types.
Those will provide the pathway, and these will be replaced by flags declaring the reference.

That is needed, cause these still imply information not innate to those entity components declared in the result types. They say, to ATTACK detection (detectors can be used for multiple reasons, and the ref needs to imply which one it is going to). The difference now, is that they'll all use the same pathways, but the kind of detector being referenced will be implied by the reference flag.