bang progress
That's the traits mostly sorted out.
I wanted to add the reference flag trait as a binding, but I've decided not to for now.
Cause the generics have to be propagated when you do that.
And it doesn't let met define a trait implementation within a trait.

The next step is macros, and I'll probably handle the reference flag implementation in those.