Bang built
So that is it built. It's messy, but it theoretically works.
I guess next step is to rebuild unit with this. I'll probably have to add some other stuff, like an always on state to bang or something. After that, I desperately want to create a diagram, because it's quite hard to picture this stuff.
I'm having to imagine it being used in a more complicated setting, to understand how the building block itself should work. The thinking is like a step removed from where it should be. When I code this, I code the building block, but when I think of it, I think of a system using that building block. It means that the block itself doesn't innately imply a lot about how it would be used.