Theoretical bug-fix
With the frame delay as a thing. It is possible to have two propagation 'waves' travelling through the tree at the same time.
This'd mean that a reset to the root's ref bangs would happen in the lifetime of a wave. That'd mean that it'd send ref bangs, then get wiped, then send ones that are possibly not reflective of the current state of the tree.
This fixes that, by having the ref-bang first check if its terminal's bang is active, before sending the ref-bang signal, as of receiving the propagating signal.