Workspace fixed
Ha...
I was having issues for around 3 days or more, and it is all because the edition field in the TOML wasn't set.
Edition, stands for Rust edition; If you don't fill it out it defaults to a 2015 version of Rust.
(Or it depends or something? In N9 it worked for a while without having to do this.)
Anyways, that was the root of my problems. Very simple...

But yeah, my issue is that it was asking me to do extern crate, but I knew from observing other people's code, that you don't have to do this, so I kept searching for what was causing it.

It's finally over though, I can get back to doing real work.