some progress
So I know how this should work now, but to actually write this as a function-like macro is difficult.
The code you see here was me trying to kind-of break the problem down, but it isn't working well.
Cause procedural macros are heavy or something, rust-analyzer struggles with them and it causes a lot of delay.

This is too learning focused rn, I'm going to move into a proto project to do some more basic things with this stuff, then I'll come back to this.