theory test
entrance_entity::query1::()::entity_binding::{}

Printed the enums of the tokens:
Ident
Punct
Punct
Ident
Punct
Punct
Group
Punct
Punct
Ident
Punct
Punct
Group

So, my understanding is that these tokens are like how it is read as english.
Step by step, information is taken in, based on how it fits into these types. Then the compiler interprets the literal text that these are, to apply the logic of the coding language.
So spaces, don't exist; They simply give us the ability to separate two idents.
Maybe, with this understanding, it could also make sense why Rust Analyzer tells you that () is void, as a return type.
Because here, we see that my optional bindings, if empty, are declared just as (). Theoretically, I could have it automatically detect if there is nothing there, so that the developer doesn't have to declare that empty ().

Maybe... It's not the same, but it reminds me of that.
But yeah, I am just progressing my understanding of this proc macro stuff. From learning, I could see how Rust could be using these tokens for its own language compilation. I don't know if it does, but it's interesting to think that it could. It does seem to utilise these rules. I suppose these rules are so simple that all languages could, or do theoretically use them, though.