Pseudocode fixed
I think this is it.
It's like, the END STEP will just have to accept a nesting level parameter I believe?
Yeah there will be complications when writing it.
I haven't thought out any of the stuff for writing the literal pathway from this, just focusing on getting it to read the thing correctly for now.