ref paths macro progress
I also considered changing to format a bit. This is what I was considering:
query()
to
query<>

I decided not to though.
Cause the query as this: query()
It's more familiar since it gets followed by a ->
Having this? query<> -> entity
Well nothing is like that in Rust, so it's kinda strange.
But it does make some sense, cause the entity bindings are mildly similar to generics in declaration and usage.
And now, to declare a tuple in there you do this: query((binding1, binding2))
But with that, it could be: query<(binding1, binding2)>

Hmm... there isn't a big difference, but I will go with: query()
I think it does make more sense.