pathway macro 3
still not finished.
Theoretically, it is kind of following the pathway format correctly now. Though, no wild-card integration yet.
Also, this is sort-of a mock-up. There're tons of errors here, but my rust-analyzer isn't showing me them.
But yeah, in theory, this is kind of it following the format correctly, I think.