pathways macro entity progress
So the Rust Analyzer was working for a while, after I commented out everything, and added them to the macro.
But it doesn't seem to want to work again, so I'm going to start a new project that is just this library, because I think rust-analyzer will work better for proc macros, if it isn't a workspace.