n11 entity progress
That's pretty much the logic for following the entity parts of the pathway done.

This thing with rust-analyzer breaking for my proc macro lib.
It seems to be caused by updating the proc macro function in the code.
As in writing anything into that function, and saving.
You can still edit it through proxy though, by editing any of the functions it calls.

You can fix it by regenerating the target.