n11 tuple ID stuff altered
If the entity field is declared in bindings, none of the tuple ID nonsense is needed.
Though, if the bindings are bound as a single tuple with the entity, then there'll need to be logic to allow the entity field from that tuple to be used as the query input.
Yeah, this way, tuple ID stuff is kept away from the bindings section, and is used as you'd expect.