n11, next steps
So, with this implementation, it is starting to transform into something that'll write the code.
So, I think the way this'll work, is that the root entity step call will return a token stream or string, and it'll use the spans and stuff that it gets from the other entity step functions.
Then for queries, it is mostly the same.