n11 entity type
So, to explain, this is the code structure that this thing is made to automate:

    // Follow reference path
    let root = to_root.entity();
    let to_control = root_q.get(root);
    let Ok(to_control) = to_control else {
        TargetFromCommandable::print_err(1);
        return;
    };
    let control = to_control.entity();
    let commandable = control_q.get(control);
    let Ok(commandable) = commandable else {
        TargetFromCommandable::print_err_descript(2, "failed at getting commandable from the entity.");
        return;
    };

The .entity() thing is a component that points to another entity.
However, a query can also search for an entity directly.
I've added a wildcard '@' which dennotes that the entity binding in question is in-fact a raw entity, which means that .entity() doesn't need to be called.

I'm also realising now, there might have to be some more logic that decides whether or not this new value will shadow the thing that pointed to it.