n11 lift wildcard progress
I'm calling it here, unfinished implementation of the wildcard that deals with handling whether or not an entity ref will be shadowed by its entity output.
By default it is shadowed, the wildcard says otherwise.
Also, tuples exist, and those cannot be lifted or shadowed, there is stuff in there for that.