n11 - basics fin
Untested.
Also un-finished, because I still have to add a writing part that'll create the unwrapping steps for the query get.

And I'm also realising now, yeah the bindings have to be bound after that step, and something placeholder has to be used beforehand.
Or maybe it is possible to do it all in a: "let Ok(bindings) = query.get() else error", from the first statement.
Will have to check.