latch rework
It's still the same system, but I've changed the 'blueprints' to not use a trait.
A trait ties the logic to the definition of a component, but the logic should be separate, so I've done that.
Now, there are simply generic functions, that act like the blueprints.
They are prefabs, in the sense that they automate the most common format that these latch systems will take.
If a dev wanted, they could make their own custom latch system, not using this stuff here.
Showing the fact, that these generic functions, are just for automation.