behaviour tree, progress
So, still working on it.
I've realised that there is nothing that deactivates a bang yet, not here and not in the design diagram.
I've decided that deactivation will be done locally, via 'fizzler' components, that will read the state terminal, to make their decision.
Since child nodes can output state to their parents, state is pushed upwards through the tree. If a latch were to be the thing to deactivate a component, that would be like distilling state down the tree.

I don't really know what the difference would be in-practice, but I'll stick to the bottom-up approach, cause that's what I naturally found myself doing.

Also with my implementation, state managing components have no reason to care about bang. Just something to keep note of.

Also, latch propagation needs to also happen whenever a state terminal is updated, on an active bang.
This isn't a thing currently.

Things are getting a little confusing but, I think this is basically it done after I make those additions. I suppose it's confusing, because I've reached a point where I can review the design of this framework I've made, and I can see that it's weird; It doesn't seem to be reflective of what behaviour trees traditionally are, I don't really know what it is. Maybe you could compare it to pure-data? The bang stuff reminds me of that, but I'm not sure if it is accurate to say that.

Hopefully, I can find some literature covering what it actually is, it feels strange to have made this, and to now not know what it really is.
I know I can use it, and how, but I don't know what it is.