1st test
and it failed.
The queries need to be disjointed.Or I could use a param set and hold the results of the 1st one in memory.
I will disjoint the queries, by making something like a BasicRelease. So it'll query for that in the children, and edit them.
There'll be another system for those, which'll apply the deactivation to their bang.