ref signature in action
It is still not exactly correct, cause the waymarks in use in this example are not signed.
That matters because, you could have waymarks pointing to things of the same kind, but of different purpose. The attack and aggro detector are examples of that; The same thing used for different purposes.
With that as fact, the waymarks need to be tied to the signature.