Bug fixing, units move now.
The units move now, there're still issues though.
I'd say the largest problem space was with the waymarks and entity reference stuff. They are spaghetti like right now. It's because I've progressed my theory on how I should handle references; There are new implementations that are built to reflect that, and there are old implementations based on old theory. It gets messy, but I'm not going to go through the struggle of updating everything.

List of bugs squashed:
The type terminal was not being issued types when units were being given commands.
The unit's behaviour did not respond to commands properly. It would get locked into states and couldn't escape them, because the orders weren't affecting behaviour.
The units were seeing eachother as enemies.
The behaviour tree was not being initialised properly.
Lots of waymarks were not in-place in the composition.
The transmission components and some transmission systems were not set-up correctly.